renaming single-results file
